+++ src/main/resources/net/uaznia/lukanus/hudson/plugins/gitparameter/GitParameterDefinition/help-defaultValue.html
@@ -1,3 +1,3 @@
 <div>
-    This value is returned when list is empty.
+    This value is returned when the list is empty or if an error occurs while retrieving data.
 </div>
