+++ src/main/resources/net/uaznia/lukanus/hudson/plugins/gitparameter/GitParameterDefinition/help-defaultValue_fr.html
@@ -1,3 +1,3 @@
 <div>
-    Valeur retournée lorsque la liste est vide.
+    Cette valeur est renvoyée lorsque la liste est vide ou en cas d'erreur lors de la récupération des données.
 </div>
