+++ src/main/resources/net/uaznia/lukanus/hudson/plugins/gitparameter/GitParameterDefinition/help-defaultValue_pl.html
@@ -1,3 +1,3 @@
 <div>
-    Ta wartość zwracana jest, kiedy lista jest pusta.
+    Ta wartość zwracana jest kiedy lista jest pusta lub jeśli wystąpi błąd podczas pobierania danych.
 </div>
