+++ src/main/resources/net/uaznia/lukanus/hudson/plugins/gitparameter/Messages.properties
@@ -1,5 +1,5 @@
 GitParameterDefinition.DisplayName=Git Parameter
-GitParameterDefinition.noRepositoryConfigured=!No Git repository configured in SCM configuration
+GitParameterDefinition.noRepositoryConfigured=No Git repository configured in SCM configuration or plugin is configured wrong
 GitParameterDefinition.invalidBranchPattern=The pattern {0} does not appear to be valid.
 GitParameterDefinition.unexpectedError=Unexpected error!
 GitParameterDefinition.lookAtLog=Please look at the Log
@@ -8,4 +8,7 @@ GitParameterDefinition.genContentsCloneDone=generateContents clone done
 GitParameterDefinition.notFindAuthorPattern=Did not find author pattern {0}
 GitParameterDefinition.invalidUseRepositoryPattern=The pattern {0} does not appear to be valid.
 GitParameterDefinition.getTag=Download tags from the repository failed
-GitParameterDefinition.error=An error occurred while download data, the default value was returned
+GitParameterDefinition.error=An error occurred while download data
+GitParameterDefinition.requiredDefaultValue=Default Value is required. Example origin/master
+GitParameterDefinition.checkConfiguration=Please check the configuration
+GitParameterDefinition.returnDefaultValue=The default value has been returned
