+++ src/main/resources/net/uaznia/lukanus/hudson/plugins/gitparameter/Messages_fr.properties
@@ -1,5 +1,5 @@
 GitParameterDefinition.DisplayName=Git Parameter
-GitParameterDefinition.noRepositoryConfigured=Aucun d\u00e9p\u00f4t git configur\u00e9 parmi les configurations scm
+GitParameterDefinition.noRepositoryConfigured=Aucun d\u00e9p\u00f4t git configur\u00e9 parmi les configurations SCM ou le plugin est mal configur\u00E9
 GitParameterDefinition.invalidBranchPattern=L'entr\u00e9e {0} semble invalide
 GitParameterDefinition.unexpectedError=Erreur inattendue !
 GitParameterDefinition.lookAtLog=Veuillez regarder les logs
@@ -8,4 +8,7 @@ GitParameterDefinition.genContentsCloneDone=Clonage termin\u00e9
 GitParameterDefinition.notFindAuthorPattern=Auncune syntaxe d'auteur trouv\u00e9e {0}
 GitParameterDefinition.invalidUseRepositoryPattern=L'entr\u00e9e {0} semble invalide
 GitParameterDefinition.getTag=La r\u00e9cup\u00e9ration des tags du d\u00e9p\u00f4t a \u00e9chou\u00e9
-GitParameterDefinition.error=Une erreur s'est produite lors du t\u00E9l\u00E9chargement des donn\u00E9es, la valeur par d\u00E9faut a \u00E9t\u00E9 renvoy\u00E9e
+GitParameterDefinition.error=Une erreur s'est produite lors du t\u00E9l\u00E9chargement des donn\u00E9es
+GitParameterDefinition.requiredDefaultValue=La valeur par d\u00E9faut est requise. Exemple origin/master
+GitParameterDefinition.checkConfiguration=S''il vous pla\u00EEt v\u00E9rifier la configuration
+GitParameterDefinition.returnDefaultValue=La valeur par d\u00E9faut a \u00E9t\u00E9 renvoy\u00E9e
