+++ src/main/resources/net/uaznia/lukanus/hudson/plugins/gitparameter/Messages_pl.properties
@@ -1,5 +1,5 @@
 GitParameterDefinition.DisplayName=Git Parameter
-GitParameterDefinition.noRepositoryConfigured=!Nie ustawiono repozytorium Git w konfiguracji SCM
+GitParameterDefinition.noRepositoryConfigured=Nie ustawiono repozytorium Git w konfiguracji SCM lub b\u0142\u0119dnie skonfigurowano wtyczk\u0119
 GitParameterDefinition.invalidBranchPattern=Wzorzec {0} nie jest prawid\u0142owym wyra\u017Ceniem regularnym.
 GitParameterDefinition.unexpectedError=Niespodziewany b\u0142\u0105d!
 GitParameterDefinition.lookAtLog=Prosz\u0119 sprawd\u017A logi
@@ -8,4 +8,7 @@ GitParameterDefinition.genContentsCloneDone=Klonowanie wykonane
 GitParameterDefinition.notFindAuthorPattern=Nie znaleziono wzorca autora
 GitParameterDefinition.invalidUseRepositoryPattern=Wzorzec {0} nie jest prawid\u0142owym wyra\u017Ceniem regularnym.
 GitParameterDefinition.getTag=Pobranie tag\u00F3w z repozytorium zako\u0144czone niepowodzeniem.
-GitParameterDefinition.error=Wyst\u0105pi\u0142 b\u0142\u0105d podczas pobierania danych, zwr\u00F3cono warto\u015B\u0107 domy\u015Blna
+GitParameterDefinition.error=Wyst\u0105pi\u0142 b\u0142\u0105d podczas pobierania danych
+GitParameterDefinition.requiredDefaultValue=Warto\u015B\u0107 domy\u015Blna jest wymagana. Przyk\u0142ad origin/master
+GitParameterDefinition.checkConfiguration=Prosz\u0119 sprawd\u017A konfiguracj\u0119
+GitParameterDefinition.returnDefaultValue=Zosta\u0142a zwr\u00F3cona warto\u015B\u0107 domy\u015Blna
